Putting everything in one JS file
