+++ client-side/Chatbot/Right.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Right Section</title>
-    <script src="../Chatbot.js" defer></script>
+    <script type="module" src="../JS/Main.js" defer></script>
     <style>
         body {
             display: flex;
