--- client-side/Login/Login.js
@@ -1,39 +1,0 @@
-import { initializeApp } from "https://www.gstatic.com/firebasejs/10.12.5/firebase-app.js";
-import { getAuth, signInWithEmailAndPassword } from "https://www.gstatic.com/firebasejs/10.12.5/firebase-auth.js";
-
-
-const firebaseConfig = {
-    apiKey: "AIzaSyC08ZzFONNjJWqK88YdXTPNL04WDDLyHG8",
-    authDomain: "adaptive-ui-webpage.firebaseapp.com",
-    projectId: "adaptive-ui-webpage",
-    storageBucket: "adaptive-ui-webpage.appspot.com",
-    messagingSenderId: "986887743719",
-    appId: "1:986887743719:web:e1eb993352fb3edd1b9a42"
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const auth = getAuth(app);
-
-const submit = document.getElementById('login-submit');
-submit.addEventListener("click", (event) => {
-    event.preventDefault();
-
-    const email = document.getElementById('login-email').value;
-    const password = document.getElementById('login-password').value;
-
-    signInWithEmailAndPassword(auth, email, password)
-        .then((userCredential) => {
-            // Signed up 
-            const user = userCredential.user;
-            console.log('Logging in  user');
-            window.location.href = "../Page.html";
-            // ...
-        })
-        .catch((error) => {
-            const errorCode = error.code;
-            const errorMessage = error.message;
-            console.log(errorMessage);
-            // ..
-        });
-})
