+++ client-side/Login/Register.html
@@ -41,6 +41,6 @@
             </div>
         </form>
 
-        <script type="module" src="Register.js" async defer></script>
+        <script type="module" src="../JS/Main.js" async defer></script>
     </body>
 </html>
